ConfigLoader detects when a provided file is not a file.  Moves ConfigException to outer scope.
