+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc-server/src/main/java/gov/cms/bfd/pipeline/rda/grpc/sink/ConfigException.java
@@ -1,0 +1,29 @@
+package gov.cms.bfd.pipeline.rda.grpc.sink;
+
+/**
+ * These exceptions are thrown by the ConfigLoader class when it encounters a requirement that is
+ * not met. For example a file doesn't exist or an integer setting isn't a valid integer string.
+ */
+public class ConfigException extends RuntimeException {
+  private final String name;
+
+  public ConfigException(String name, String message) {
+    super(message);
+    this.name = name;
+  }
+
+  public ConfigException(String name, String message, Throwable cause) {
+    super(message, cause);
+    this.name = name;
+  }
+
+  /** @return The name of the configuration option that had an invalid value. */
+  public String getName() {
+    return name;
+  }
+
+  @Override
+  public String toString() {
+    return String.format("invalid option: name='%s' message='%s'", name, getMessage());
+  }
+}
